refactor: give log transactions direct ownership
