+++ plugins/codex-autoresearch/lib/git-private-state.ts
@@ -2,8 +2,9 @@ import fs from "node:fs";
 import path from "node:path";
 
 import { isPathInside } from "./path-containment.js";
-import { runProcess } from "./runner.js";
+import { runProcess, type ProcessRunResult } from "./runner.js";
 import { displayGitPath, parsePorcelainV1Z } from "./git-paths.js";
+import { pathExists } from "./session-core.js";
 
 function hasGitMarker(cwd: string): boolean {
   let current = path.resolve(cwd);
@@ -15,22 +16,22 @@ function hasGitMarker(cwd: string): boolean {
   }
 }
 
-async function git(args: string[], cwd: string) {
+export async function runGit(args: string[], cwd: string) {
   return await runProcess("git", args, { cwd, maxOutputBytes: 16 * 1024 * 1024 });
 }
 
-function gitOutput(result: { stderr: string; stdout: string }, fallback: string): string {
+export function gitOutput(result: { stderr: string; stdout: string }, fallback: string): string {
   return (result.stderr || result.stdout || fallback).trim();
 }
 
 export async function insideGitRepo(cwd: string): Promise<boolean> {
   if (!hasGitMarker(cwd)) return false;
-  const result = await git(["rev-parse", "--is-inside-work-tree"], cwd);
+  const result = await runGit(["rev-parse", "--is-inside-work-tree"], cwd);
   return result.code === 0 && result.stdout.trim() === "true";
 }
 
 export async function gitPrivatePath(cwd: string, relativePath: string): Promise<string> {
-  const result = await git(["rev-parse", "--git-path", relativePath], cwd);
+  const result = await runGit(["rev-parse", "--git-path", relativePath], cwd);
   if (result.code !== 0) {
     throw new Error(`Git path lookup failed: ${gitOutput(result, "unknown error")}`);
   }
@@ -39,7 +40,7 @@ export async function gitPrivatePath(cwd: string, relativePath: string): Promise
 }
 
 export async function gitPrivateRoot(cwd: string): Promise<string> {
-  const result = await git(["rev-parse", "--git-dir"], cwd);
+  const result = await runGit(["rev-parse", "--git-dir"], cwd);
   if (result.code !== 0) {
     throw new Error(`Git directory lookup failed: ${gitOutput(result, "unknown error")}`);
   }
@@ -58,7 +59,7 @@ export async function privateStateWriteRoot(workDir: string, target: string): Pr
 
 export async function gitDirtyPathDetails(cwd: string) {
   if (!(await insideGitRepo(cwd))) return [];
-  const result = await git(["status", "--porcelain=v1", "-z", "-uall"], cwd);
+  const result = await runGit(["status", "--porcelain=v1", "-z", "-uall"], cwd);
   if (result.code !== 0) {
     throw new Error(`Git status failed: ${gitOutput(result, "unknown error")}`);
   }
@@ -74,4 +75,101 @@ export async function gitDirtyPathDetails(cwd: string) {
       raw: `${entry.status} ${displayGitPath(gitPath)}`,
     })),
   );
+}
+
+export async function gitStatusShort(cwd: string): Promise<string> {
+  const result = await runGit(["status", "--porcelain=v1", "-z", "-uall"], cwd);
+  if (result.code !== 0) {
+    throw new Error(`Git status failed: ${gitOutput(result, "unknown error")}`);
+  }
+  assertCompleteGitPathOutput(result);
+  return result.stdout;
+}
+
+export async function shortHead(cwd: string): Promise<string> {
+  const result = await runGit(["rev-parse", "--short=7", "HEAD"], cwd);
+  return result.code === 0 ? result.stdout.trim() : "";
+}
+
+export async function resolveCommitRef(cwd: string, commit: unknown): Promise<string> {
+  const value = String(commit || "").trim();
+  if (!value) throw new Error("commit is required");
+  const result = await runGit(["rev-parse", "--verify", `${value}^{commit}`], cwd);
+  if (result.code !== 0) {
+    throw new Error(`Git commit could not be resolved: ${gitOutput(result, value)}`);
+  }
+  return result.stdout.trim();
+}
+
+export async function hasStagedChanges(cwd: string): Promise<boolean> {
+  const result = await runGit(["diff", "--cached", "--quiet"], cwd);
+  return result.code === 1;
+}
+
+export async function hasStagedChangesInPaths(cwd: string, paths: string[]): Promise<boolean> {
+  const result = await runGit(
+    ["--literal-pathspecs", "diff", "--cached", "--quiet", "--", ...paths],
+    cwd,
+  );
+  return result.code === 1;
+}
+
+export async function assertNoGitIndexLock(
+  workDir: string,
+  phase = "git operation",
+): Promise<void> {
+  const lockPath = await gitPrivatePath(workDir, "index.lock");
+  if (!(await pathExists(lockPath))) return;
+  throw new Error(await gitIndexLockMessage(workDir, lockPath, phase, false));
+}
+
+export function gitIndexLockFailure(result: { stderr: string; stdout: string }): boolean {
+  return /index\.lock|another git process|Unable to create/i.test(gitOutput(result, ""));
+}
+
+export async function gitIndexLockMessage(
+  workDir: string,
+  lockPath: string,
+  phase: string,
+  stagedMayHaveChanged: boolean,
+): Promise<string> {
+  const liveGit = await liveGitProcessSummary(workDir);
+  return [
+    `Git index lock blocked ${phase}: ${lockPath}.`,
+    `Live git process check: ${liveGit}.`,
+    stagedMayHaveChanged
+      ? "Autoresearch could not prove whether staging partially changed; inspect git status before retrying."
+      : "Autoresearch has not staged or committed anything for this log attempt.",
+    "Wait for active Git commands to finish, then retry. If no Git process is active, remove the index.lock file and rerun the exact log command.",
+  ].join(" ");
+}
+
+function assertCompleteGitPathOutput(result: ProcessRunResult): void {
+  if (result.stdoutTruncated) {
+    throw new Error(
+      "Git path output exceeded the capture limit; refusing an incomplete trust check.",
+    );
+  }
+}
+
+async function liveGitProcessSummary(workDir: string): Promise<string> {
+  try {
+    const result =
+      process.platform === "win32"
+        ? await runProcess(
+            "powershell",
+            [
+              "-NoProfile",
+              "-Command",
+              "Get-Process git -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Id",
+            ],
+            { cwd: workDir, timeoutSeconds: 2 },
+          )
+        : await runProcess("pgrep", ["-fl", "git"], { cwd: workDir, timeoutSeconds: 2 });
+    const outputText = `${result.stdout || ""}${result.stderr ? `\n${result.stderr}` : ""}`.trim();
+    if (!outputText) return "no live git process found";
+    return outputText.split(/\r?\n/).slice(0, 5).join(", ");
+  } catch (error) {
+    return `process check unavailable (${error instanceof Error ? error.message : String(error)})`;
+  }
 }
